chore: Moved `suggester.ts` into `./suggester`
