+++ web/src/Web.App/AssetSrc/ts/index.d.ts
@@ -1,0 +1,5 @@
+interface ApiError {
+  error?: Error;
+}
+
+type ApiResponse<T> = T[] | ApiError;
