--- web/src/Web.App/AssetSrc/ts/index.ts
@@ -1,3 +1,0 @@
-import { suggester } from "./suggester.ts";
-
-export { suggester };
