+++ web/src/Web.App/AssetSrc/ts/main.ts
@@ -1,0 +1,3 @@
+import { suggester } from "./suggester/component.ts";
+
+export { suggester };
