+++ web/src/Web.App/AssetSrc/ts/suggester/index.d.ts
@@ -1,0 +1,4 @@
+interface SuggestResult<T> {
+  text: string;
+  document: T;
+}
